Move `coreCommands` out of `/controller/` (#141174)
