+++ src/vs/editor/browser/coreCommands.ts
